[REFACTOR]Dialog들 분리
